Fixed the internal txn ids for the protodtxncoordinator to be 64-bit. I think we're in business!!!
git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/protorpc-messaging-branch@1641 7a8f3d5b-8c5f-44cb-9125-2414a256df87

+++ src/dtxn/dtxn/messages.py
@@ -1,4 +1,5 @@
 #!/usr/bin/python
+# -*- coding: utf-8 -*-
 
 from stupidcompiler import *
 
@@ -8,7 +9,7 @@ from stupidcompiler import *
 request = MessageDefinition("Fragment", "Contains a transaction fragment to be processed by the execution engine.")
 # TODO: Add client ids so this id is unique across the entire DB?
 request.addField(INT32, "client_id", 0, "Unique identifier for the client sending the transaction.")
-request.addField(INT32, "id", 0, "Unique identifier for this transaction. Used to identify messages that belong to a single transaction.")
+request.addField(INT64, "id", 0, "Unique identifier for this transaction. Used to identify messages that belong to a single transaction.")
 request.addField(BOOL, "multiple_partitions", False, "True if this transaction spans multiple partitions.")
 request.addField(BOOL, "last_fragment", True, "True if this is the last fragment of this transaction for this partition.")
 # TODO: Rename this to command? or fragment?
@@ -16,7 +17,7 @@ request.addField(STRING, "transaction", None, "The actual transaction to be proc
 request.addField(STRING, "payload", None, "Fragment messsage payload")
 
 response = MessageDefinition("FragmentResponse", "The response from the partition.")
-response.addField(INT32, "id", 0, "Identifies the transaction that this decision refers to.")
+response.addField(INT64, "id", 0, "Identifies the transaction that this decision refers to.")
 # TODO: Change this to an enum
 response.addField(INT32, "status", -1, "Contains the ExecutionEngine::Status for this transaction.")
 # Rename this to output? Might be more clear.
@@ -26,7 +27,7 @@ response.addField(INT32, "dependency", -1, "ID of a previous transaction that th
 
 commit_decision = MessageDefinition("CommitDecision", "Instructs a replica to commit or abort a transaction.")
 commit_decision.addField(INT32, "client_id", 0, "Unique identifier for the client sending the transaction.")
-commit_decision.addField(INT32, "id", 0, "Identifies the transaction that this decision refers to.")
+commit_decision.addField(INT64, "id", 0, "Identifies the transaction that this decision refers to.")
 commit_decision.addField(BOOL, "commit", False, "True if the transaction should commit.")
 commit_decision.addField(STRING, "payload", None, "Commit messsage payload")
