+++ tests/frontend/org/voltdb/TestTransactionState.java
@@ -86,7 +86,7 @@ public class TestTransactionState extends BaseTestCase {
             ftasks = plan.getFragmentTaskMessages(TXN_ID, CLIENT_HANDLE);
             assertFalse(ftasks.isEmpty());
         }
-        this.ts = new TransactionState(executor, TXN_ID, TXN_ID.intValue(), LOCAL_PARTITION, CLIENT_HANDLE, EXEC_LOCAL);
+        this.ts = new TransactionState(executor, TXN_ID, TXN_ID, LOCAL_PARTITION, CLIENT_HANDLE, EXEC_LOCAL);
         assertNotNull(this.ts);
         assertEquals(TransactionState.RoundState.NULL, this.ts.getCurrentRoundState());
     }
